Refactored code

+++ src/main/java/com/groupthree/shopsphere/repository/UserRepository.java
@@ -1,8 +1,8 @@
 package com.groupthree.shopsphere.repository;
-// Interface for interacting with users data.
+
 import com.groupthree.shopsphere.models.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User,Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     User findByEmail(String email);
 }
